requirements.txt
streamlit
pillow
python-docx
requests
